+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # @gumbee/structured
 
+## 2.0.0
+
+### Major Changes
+
+- initial release of structured package
+
+### Minor Changes
+
+- allow inline rule and description definition and fix utility types in typescript def
+
 ## 1.0.0
 
 ### Major Changes